+++ lib/flutter_gyw.dart
@@ -7,6 +7,7 @@ export 'src/bt_manager.dart' show GYWBtManager;
 export 'src/drawings.dart'
     show GYWDrawing, TextDrawing, BlankScreen, IconDrawing;
 export 'src/exceptions.dart' show GYWException, GYWStatusException;
+export 'src/firmware_version.dart' show FirmwareVersion;
 export 'src/fonts.dart' show GYWFont;
 export 'src/icons.dart' show GYWIcon;
 export 'src/screen.dart' show GYWScreenParameters;