+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * [ADD] Convert WhiteScreen into BlankScreen with a color parameter
 * [FIX] Upgrade flutter_blue_plus version
 * [ADD] Add a method to enable or disable the display backlight
+* [ADD] Add a method to read the firmware version
 
 ## 1.1.1
 * [IMP] Integrate font in text control instruction